requirements.txt
altgraph==0.10.2
astroid==1.5.3
backports.functools-lru-cache==1.4
bdist-mpkg==0.5.0
beautifulsoup4==4.6.0
bonjour-py==0.3
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
diff-match-patch==20121119
Django==1.11.7
django-import-export==0.5.1
djangorestframework==3.7.1
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.6
isort==4.2.15
jdcal==1.3
lazy-object-proxy==1.3.1
listorm==0.2.12
macholib==1.5.1
matplotlib==1.3.1
mccabe==0.6.1
modulegraph==0.10.4
numpy==1.8.0rc1
odfpy==1.3.5
openpyxl==2.4.9
py2app==0.7.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
PyYAML==3.12
requests==2.18.4
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.4.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
wrapt==1.10.11
xattr==0.6.4
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zope.interface==4.1.1
